requirements.txt
pillow~=11.1.0
torch~=2.6.0
transformers~=4.49.0
openai~=0.28.0
numpy~=2.2.3
faiss-cpu~=1.10.0
pdf2image~=1.17.0
#make sure to install Poppler on the system (apt-get install poppler-utils)
